# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PhoAmThuc API",
    "description": "Auto-generated OpenAPI spec from Express routes.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/v1/sync/manifest": {
      "get": {
        "description": "Return current content version and sync metadata.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncManifestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Sync"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "get_api_v1_sync_manifest",
        "summary": "Get sync manifest"
      }
    },
    "/api/v1/sync/full": {
      "get": {
        "description": "Return full POI and tour data for bootstrap or fallback sync.",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncFullResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Sync"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "get_api_v1_sync_full",
        "summary": "Get full sync payload"
      }
    },
    "/api/v1/sync/incremental": {
      "post": {
        "description": "Return delta changes from a client content version.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncIncrementalSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncIncrementalConflictResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fromVersion"
                ],
                "properties": {
                  "fromVersion": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Sync"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "post_api_v1_sync_incremental",
        "summary": "Get incremental sync payload"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Validation failed"
              }
            }
          }
        }
      },
      "LanguageMap": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      },
...
      "PublicPoiItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/LanguageMap"
          },
          "description": {
            "$ref": "#/components/schemas/LanguageMap"
          },
          "audioUrls": {
            "$ref": "#/components/schemas/LanguageMap"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PublicTourItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/LanguageMap"
          },
          "description": {
            "$ref": "#/components/schemas/LanguageMap"
          },
          "duration": {
            "type": "integer"
          },
          "poiIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SyncManifestResponse": {
        "type": "object",
        "properties": {
          "contentVersion": {
            "type": "integer"
          },
          "totalPois": {
            "type": "integer"
          },
          "totalTours": {
            "type": "integer"
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "checksum": {
            "type": "string"
          }
        }
      },
      "SyncFullResponse": {
        "type": "object",
        "properties": {
          "contentVersion": {
            "type": "integer"
          },
          "needsSync": {
            "type": "boolean"
          },
          "pois": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicPoiItem"
            }
          },
          "tours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicTourItem"
            }
          }
        }
      },
      "SyncIncrementalSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "SyncIncrementalConflictResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "error": {
            "type": "object"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
    }
  },
...
}